requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
sentence-transformers==3.3.1
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
trafilatura>=1.6.0
transformers==4.55.0
typing_extensions==4.14.1
urllib3==2.5.0
python-telegram-bot==13.15
streamlit
plotly
yfinance
pytz
